requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.12.0=pypi_0
addict=2.4.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
apex=0.1=pypi_0
argon2-cffi=20.1.0=py38h27cfd23_1
argparse=1.4.0=pypi_0
async_generator=1.10=pyhd3eb1b0_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
batchgenerators=0.24=pypi_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
ca-certificates=2021.9.30=h06a4308_1
cachetools=4.2.1=pypi_0
certifi=2021.10.8=py38h06a4308_0
cffi=1.14.6=py38h400218f_0
chardet=4.0.0=pypi_0
click=8.0.3=pypi_0
cloudpickle=1.6.0=pypi_0
color-transfer=0.1=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=pypi_0
cython=0.29.22=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.4.1=py38h295c915_0
decorator=4.4.2=pypi_0
defusedxml=0.7.1=pyhd3eb1b0_0
dicom2nifti=2.3.3=pypi_0
entrypoints=0.3=py38_0
et-xmlfile=1.1.0=pypi_0
expat=2.4.1=h2531618_2
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
future=0.18.2=pypi_0
fvcore=0.1.5.post20210410=pypi_0
glib=2.69.1=h5202010_0
google-auth=1.28.1=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
grpcio=1.37.0=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.6.0=pypi_0
icu=58.2=he6710b0_3
idna=2.10=pypi_0
imageio=2.9.0=pypi_0
imbalanced-learn=0.8.0=pypi_0
imblearn=0.0=pypi_0
importlib-metadata=4.8.1=py38h06a4308_0
importlib_metadata=4.8.1=hd3eb1b0_0
intel-openmp=2020.2=254
iopath=0.1.7=pypi_0
ipykernel=6.4.1=py38h06a4308_1
ipython=7.27.0=py38hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.0=py38h06a4308_1
jinja2=3.0.1=pyhd3eb1b0_0
joblib=1.0.1=pypi_0
jpeg=9b=h024ee3a_2
jsonpatch=1.32=pypi_0
jsonpointer=2.2=pypi_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter=1.0.0=py38_7
jupyter_client=7.0.1=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.8.1=py38h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.1=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h7f8727e_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
linecache2=1.0.0=pypi_0
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=pypi_0
markupsafe=2.0.1=py38h27cfd23_0
matplotlib=3.4.1=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
medpy=0.4.0=pypi_0
mistune=0.8.4=py38h7b6447c_1000
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.3.0=py38h54f3939_0
mkl_random=1.1.1=py38h0573a6f_0
mock=4.0.3=pypi_0
mxnet=1.8.0.post0=pypi_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py38h06a4308_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.5.1=pypi_0
nibabel=3.2.2=pypi_0
ninja=1.10.2=py38hff7bd54_0
nnunet=1.7.0=pypi_0
notebook=6.4.3=py38h06a4308_0
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
omegaconf=2.1.0.dev22=pypi_0
opencv-python=4.2.0.34=pypi_0
openpyxl=3.1.2=pypi_0
openslide-python=1.1.2=pypi_0
openssl=1.1.1l=h7f8727e_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.2.4=py38h2531618_0
pandocfilters=1.4.3=py38h06a4308_1
parso=0.8.2=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py38he98fc37_0
pip=21.0.1=py38h06a4308_0
polygon3=3.0.9.1=pypi_0
portalocker=2.3.0=pypi_0
prometheus_client=0.11.0=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.15.8=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.2=pypi_0
pycparser=2.20=py_2
pydensecrf=1.0rc2=pypi_0
pydicom=2.3.0=pypi_0
pydot=1.4.2=pypi_0
pygments=2.10.0=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.17.3=py38h7b6447c_0
python=3.8.8=hdb3f193_4
python-dateutil=2.8.1=pyhd3eb1b0_0
python-graphviz=0.8.4=pypi_0
pytorch=1.6.0=py3.8_cuda10.1.243_cudnn7.6.3_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.2.1=py38h295c915_1
qt=5.9.7=h5867ecd_1
qtconsole=5.1.1=pyhd3eb1b0_0
qtpy=1.10.0=pyhd3eb1b0_0
rapidfuzz=1.4.1=pypi_0
readline=8.1=h27cfd23_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scikit-image=0.18.1=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.6.2=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=52.0.0=py38h06a4308_0
shapely=1.8a1=pypi_0
simpleitk=2.1.1.2=pypi_0
sip=4.19.13=py38he6710b0_0
six=1.15.0=py38h06a4308_0
sklearn=0.0=pypi_0
sqlite=3.35.4=hdfb4753_0
tabulate=0.8.9=pypi_0
tensorboard=2.4.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorboardx=2.3=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.4=py38h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.1.0=pypi_0
tifffile=2021.4.8=pypi_0
tk=8.6.10=hbc83047_0
torchfile=0.1.0=pypi_0
torchnet=0.0.4=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.7.0=py38_cu101
tornado=6.1=py38h27cfd23_0
tqdm=4.60.0=pypi_0
traceback2=1.4.0=pypi_0
traitlets=5.1.0=pyhd3eb1b0_0
unittest2=1.1.0=pypi_0
urllib3=1.26.4=pypi_0
visdom=0.1.8.9=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
websocket-client=1.2.1=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pypi_0
widgetsnbextension=3.5.1=py38_0
xlrd=2.0.1=pypi_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
